+++ saleor/graphql/checkout/mutations/checkout_add_promo_code.py
@@ -9,13 +9,12 @@ from ....checkout.fetch import (
     update_delivery_method_lists_for_checkout_info,
 )
 from ....checkout.utils import add_promo_code_to_checkout
-from ...core.descriptions import DEPRECATED_IN_3X_INPUT
+from ...core.descriptions import ADDED_IN_34, DEPRECATED_IN_3X_INPUT
 from ...core.mutations import BaseMutation
 from ...core.scalars import UUID
 from ...core.types import CheckoutError
-from ...core.validators import validate_one_of_args_is_in_mutation
 from ..types import Checkout
-from .utils import get_checkout_by_token, update_checkout_shipping_method_if_invalid
+from .utils import get_checkout, update_checkout_shipping_method_if_invalid
 
 
 class CheckoutAddPromoCode(BaseMutation):
@@ -24,13 +23,20 @@ class CheckoutAddPromoCode(BaseMutation):
     )
 
     class Arguments:
+        id = graphene.ID(
+            description="The checkout's ID." + ADDED_IN_34,
+            required=False,
+        )
         checkout_id = graphene.ID(
             description=(
-                f"The ID of the checkout. {DEPRECATED_IN_3X_INPUT} Use token instead."
+                f"The ID of the checkout.{DEPRECATED_IN_3X_INPUT} Use `id` instead."
             ),
             required=False,
         )
-        token = UUID(description="Checkout token.", required=False)
+        token = UUID(
+            description=f"Checkout token.{DEPRECATED_IN_3X_INPUT} Use `id` instead.",
+            required=False,
+        )
         promo_code = graphene.String(
             description="Gift card code or voucher code.", required=True
         )
@@ -41,19 +47,17 @@ class CheckoutAddPromoCode(BaseMutation):
         error_type_field = "checkout_errors"
 
     @classmethod
-    def perform_mutation(cls, _root, info, promo_code, checkout_id=None, token=None):
-        # DEPRECATED
-        validate_one_of_args_is_in_mutation(
-            CheckoutErrorCode, "checkout_id", checkout_id, "token", token
+    def perform_mutation(
+        cls, _root, info, promo_code, checkout_id=None, token=None, id=None
+    ):
+        checkout = get_checkout(
+            cls,
+            info,
+            checkout_id=checkout_id,
+            token=token,
+            id=id,
+            error_class=CheckoutErrorCode,
         )
-
-        if token:
-            checkout = get_checkout_by_token(token)
-        # DEPRECATED
-        else:
-            checkout = cls.get_node_or_error(
-                info, checkout_id or token, only_type=Checkout, field="checkout_id"
-            )
 
         validate_checkout_email(checkout)