+++ saleor/graphql/checkout/tests/test_checkout_promo_codes.py
@@ -1,4 +1,3 @@
-import uuid
 from datetime import date, timedelta
 from decimal import Decimal
 from unittest import mock
@@ -19,6 +18,7 @@ from ....core.taxes import TaxedMoney
 from ....discount import DiscountInfo, VoucherType
 from ....plugins.manager import get_plugins_manager
 from ....warehouse.models import Stock
+from ...core.utils import to_global_id_or_none
 from ...tests.utils import get_graphql_content
 from .test_checkout import MUTATION_CHECKOUT_SHIPPING_ADDRESS_UPDATE
 from .test_checkout_lines import MUTATION_CHECKOUT_LINE_DELETE
@@ -47,7 +47,7 @@ def test_checkout_lines_delete_with_not_applicable_voucher(
     line = checkout_with_item.lines.first()
 
     line_id = graphene.Node.to_global_id("CheckoutLine", line.pk)
-    variables = {"token": checkout_with_item.token, "lineId": line_id}
+    variables = {"id": to_global_id_or_none(checkout_with_item), "lineId": line_id}
     response = user_api_client.post_graphql(MUTATION_CHECKOUT_LINE_DELETE, variables)
     content = get_graphql_content(response)
 
@@ -84,7 +84,10 @@ def test_checkout_shipping_address_update_with_not_applicable_voucher(
     assert checkout_with_item.voucher_code == voucher.code
 
     new_address = graphql_address_data
-    variables = {"token": checkout_with_item.token, "shippingAddress": new_address}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "shippingAddress": new_address,
+    }
     response = user_api_client.post_graphql(
         MUTATION_CHECKOUT_SHIPPING_ADDRESS_UPDATE, variables
     )
@@ -108,8 +111,8 @@ def test_checkout_totals_use_discounts(
     sale.products.add(product)
 
     query = """
-    query getCheckout($token: UUID) {
-        checkout(token: $token) {
+    query getCheckout($id: ID) {
+        checkout(id: $id) {
             lines {
                 totalPrice {
                     gross {
@@ -131,7 +134,7 @@ def test_checkout_totals_use_discounts(
     }
     """
 
-    variables = {"token": str(checkout.token)}
+    variables = {"id": to_global_id_or_none(checkout)}
     response = api_client.post_graphql(query, variables)
     content = get_graphql_content(response)
     data = content["data"]["checkout"]
@@ -174,8 +177,8 @@ def test_checkout_totals_use_discounts(
 
 
 QUERY_GET_CHECKOUT_GIFT_CARD_CODES = """
-query getCheckout($token: UUID!) {
-  checkout(token: $token) {
+query getCheckout($id: ID) {
+  checkout(id: $id) {
     token
     giftCards {
       last4CodeChars
@@ -190,7 +193,7 @@ query getCheckout($token: UUID!) {
 
 def test_checkout_get_gift_card_code(user_api_client, checkout_with_gift_card):
     gift_card = checkout_with_gift_card.gift_cards.first()
-    variables = {"token": str(checkout_with_gift_card.token)}
+    variables = {"id": to_global_id_or_none(checkout_with_gift_card)}
     response = user_api_client.post_graphql(
         QUERY_GET_CHECKOUT_GIFT_CARD_CODES, variables
     )
@@ -207,7 +210,7 @@ def test_checkout_get_gift_card_codes(
     checkout_with_gift_card.save()
     gift_card_first = checkout_with_gift_card.gift_cards.first()
     gift_card_last = checkout_with_gift_card.gift_cards.last()
-    variables = {"token": str(checkout_with_gift_card.token)}
+    variables = {"id": to_global_id_or_none(checkout_with_gift_card)}
     response = user_api_client.post_graphql(
         QUERY_GET_CHECKOUT_GIFT_CARD_CODES, variables
     )
@@ -220,7 +223,7 @@ def test_checkout_get_gift_card_codes(
 
 
 def test_checkout_get_gift_card_code_without_gift_card(user_api_client, checkout):
-    variables = {"token": str(checkout.token)}
+    variables = {"id": to_global_id_or_none(checkout)}
     response = user_api_client.post_graphql(
         QUERY_GET_CHECKOUT_GIFT_CARD_CODES, variables
     )
@@ -230,9 +233,9 @@ def test_checkout_get_gift_card_code_without_gift_card(user_api_client, checkout
 
 
 MUTATION_CHECKOUT_ADD_PROMO_CODE = """
-    mutation($token: UUID, $promoCode: String!) {
+    mutation($id: ID, $promoCode: String!) {
         checkoutAddPromoCode(
-            token: $token, promoCode: $promoCode) {
+            id: $id, promoCode: $promoCode) {
             errors {
                 field
                 message
@@ -277,7 +280,10 @@ def _mutate_checkout_add_promo_code(client, variables):
 
 def test_checkout_add_voucher_for_entire_order(api_client, checkout_with_item, voucher):
     # given
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     assert voucher.type == VoucherType.ENTIRE_ORDER
     manager = get_plugins_manager()
     lines, _ = fetch_checkout_lines(checkout_with_item)
@@ -305,7 +311,10 @@ def test_checkout_add_voucher_for_entire_order(api_client, checkout_with_item, v
 
 
 def test_checkout_add_voucher_code_by_token(api_client, checkout_with_item, voucher):
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -345,7 +354,10 @@ def test_checkout_add_voucher_code_by_token_with_external_shipment(
     set_external_shipping_id(checkout, external_shipping_method_id)
     checkout.save(update_fields=["shipping_address", "private_metadata"])
 
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -374,7 +386,10 @@ def test_checkout_add_voucher_code_with_display_gross_prices(
         ),
     )
 
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -404,7 +419,10 @@ def test_checkout_add_voucher_code_without_display_gross_prices(
         ),
     )
 
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert data["errors"][0]["code"] == CheckoutErrorCode.VOUCHER_NOT_APPLICABLE.name
@@ -417,7 +435,10 @@ def test_checkout_add_voucher_code_without_display_gross_prices(
 def test_checkout_add_voucher_code_variant_unavailable(
     api_client, checkout_with_item, voucher
 ):
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     checkout_with_item.lines.first().variant.channel_listings.filter(
         channel=checkout_with_item.channel
     ).delete()
@@ -450,7 +471,7 @@ def test_checkout_add_voucher_code_checkout_with_sale(
     previous_checkout_last_change = checkout_with_item.last_change
 
     variables = {
-        "token": checkout_with_item.token,
+        "id": to_global_id_or_none(checkout_with_item),
         "promoCode": voucher_percentage.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -489,7 +510,10 @@ def test_checkout_add_specific_product_voucher_code_checkout_with_sale(
     )
 
     assert subtotal > subtotal_discounted
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -544,7 +568,10 @@ def test_checkout_add_products_voucher_code_checkout_with_sale(
         discounts=[discount_info],
     )
     assert subtotal > subtotal_discounted
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -596,7 +623,10 @@ def test_checkout_add_collection_voucher_code_checkout_with_sale(
         discounts=[discount_info],
     )
     assert subtotal > subtotal_discounted
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -649,7 +679,10 @@ def test_checkout_add_category_code_checkout_with_sale(
         discounts=[discount_info],
     )
     assert subtotal > subtotal_discounted
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -674,7 +707,7 @@ def test_checkout_add_voucher_code_not_applicable_voucher(
     api_client, checkout_with_item, voucher_with_high_min_spent_amount
 ):
     variables = {
-        "token": checkout_with_item.token,
+        "id": to_global_id_or_none(checkout_with_item),
         "promoCode": voucher_with_high_min_spent_amount.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -687,7 +720,7 @@ def test_checkout_add_voucher_code_not_assigned_to_channel(
     api_client, checkout_with_item, voucher_without_channel
 ):
     variables = {
-        "token": checkout_with_item.token,
+        "id": to_global_id_or_none(checkout_with_item),
         "promoCode": voucher_without_channel.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -698,7 +731,10 @@ def test_checkout_add_voucher_code_not_assigned_to_channel(
 
 def test_checkout_add_gift_card_code(api_client, checkout_with_item, gift_card):
     gift_card_id = graphene.Node.to_global_id("GiftCard", gift_card.pk)
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -713,7 +749,7 @@ def test_checkout_add_many_gift_card_code(
     assert checkout_with_gift_card.gift_cards.count() > 0
     gift_card_id = graphene.Node.to_global_id("GiftCard", gift_card_created_by_staff.pk)
     variables = {
-        "token": checkout_with_gift_card.token,
+        "id": to_global_id_or_none(checkout_with_gift_card),
         "promoCode": gift_card_created_by_staff.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -731,7 +767,10 @@ def test_checkout_add_inactive_gift_card_code(
     gift_card.is_active = False
     gift_card.save(update_fields=["is_active"])
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(staff_api_client, variables)
@@ -750,7 +789,10 @@ def test_checkout_add_expired_gift_card_code(
     gift_card.expiry_date = date.today() - timedelta(days=10)
     gift_card.save(update_fields=["expiry_date"])
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(staff_api_client, variables)
@@ -771,7 +813,10 @@ def test_checkout_add_used_gift_card_code(
 
     gift_card_id = graphene.Node.to_global_id("GiftCard", gift_card_used.pk)
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card_used.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card_used.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(staff_api_client, variables)
@@ -793,7 +838,10 @@ def test_checkout_add_used_gift_card_code_invalid_user(
     checkout_with_item.user = staff_user
     assert gift_card_used.used_by_email != checkout_with_item.user.email
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card_used.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card_used.code,
+    }
 
     # when
     data = _mutate_checkout_add_promo_code(staff_api_client, variables)
@@ -817,7 +865,10 @@ def test_checkout_get_total_with_gift_card(api_client, checkout_with_item, gift_
     )
     total_with_gift_card = taxed_total.gross.amount - gift_card.current_balance_amount
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -846,7 +897,7 @@ def test_checkout_get_total_with_many_gift_card(
 
     assert checkout_with_gift_card.gift_cards.count() > 0
     variables = {
-        "token": checkout_with_gift_card.token,
+        "id": to_global_id_or_none(checkout_with_gift_card),
         "promoCode": gift_card_created_by_staff.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -862,7 +913,10 @@ def test_checkout_get_total_with_more_money_on_gift_card(
     checkout_with_item.email = gift_card_used.used_by_email
     checkout_with_item.save(update_fields=["email"])
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card_used.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card_used.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -875,7 +929,10 @@ def test_checkout_add_same_gift_card_code(api_client, checkout_with_gift_card):
     gift_card = checkout_with_gift_card.gift_cards.first()
     gift_card_id = graphene.Node.to_global_id("GiftCard", gift_card.pk)
     gift_card_count = checkout_with_gift_card.gift_cards.count()
-    variables = {"token": checkout_with_gift_card.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_gift_card),
+        "promoCode": gift_card.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert not data["errors"]
@@ -891,7 +948,10 @@ def test_checkout_add_gift_card_code_in_active_gift_card(
     gift_card.is_active = False
     gift_card.save()
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert data["errors"]
@@ -904,23 +964,30 @@ def test_checkout_add_gift_card_code_in_expired_gift_card(
     gift_card.expiry_date = date.today() - timedelta(days=1)
     gift_card.save()
 
-    variables = {"token": checkout_with_item.token, "promoCode": gift_card.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": gift_card.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert data["errors"]
     assert data["errors"][0]["field"] == "promoCode"
 
 
-def test_checkout_add_promo_code_invalid_checkout(api_client, voucher):
-    variables = {"token": uuid.uuid4(), "promoCode": voucher.code}
+def test_checkout_add_promo_code_invalid_checkout(api_client, voucher, checkout):
+    variables = {"id": to_global_id_or_none(checkout), "promoCode": voucher.code}
+    checkout.delete()
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert data["errors"]
-    assert data["errors"][0]["field"] == "token"
+    assert data["errors"][0]["field"] == "id"
 
 
 def test_checkout_add_promo_code_invalid_promo_code(api_client, checkout_with_item):
-    variables = {"token": checkout_with_item.token, "promoCode": "unexisting_code"}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": "unexisting_code",
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert data["errors"]
@@ -956,7 +1023,7 @@ def test_checkout_add_promo_code_invalidate_shipping_method(
     add_variant_to_checkout(checkout_info, variant, 5)
 
     # Apply voucher
-    variables = {"token": checkout.token, "promoCode": voucher.code}
+    variables = {"id": to_global_id_or_none(checkout), "promoCode": voucher.code}
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     shipping_method_id = graphene.Node.to_global_id(
@@ -972,7 +1039,10 @@ def test_checkout_add_promo_code_no_checkout_email(
     checkout_with_item.email = None
     checkout_with_item.save(update_fields=["email"])
 
-    variables = {"token": checkout_with_item.token, "promoCode": voucher.code}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": voucher.code,
+    }
     data = _mutate_checkout_add_promo_code(api_client, variables)
 
     assert data["errors"]
@@ -1014,7 +1084,7 @@ def test_checkout_add_free_shipping_voucher_do_not_invalidate_shipping_method(
     shipping_listing.save(update_fields=["price_amount", "minimum_order_price_amount"])
 
     variables = {
-        "token": checkout_with_item.token,
+        "id": to_global_id_or_none(checkout_with_item),
         "promoCode": voucher_free_shipping.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -1065,7 +1135,7 @@ def test_checkout_add_shipping_voucher_do_not_invalidate_shipping_method(
     voucher_shipping_type.save(update_fields=["countries"])
 
     variables = {
-        "token": checkout_with_item.token,
+        "id": to_global_id_or_none(checkout_with_item),
         "promoCode": voucher_shipping_type.code,
     }
     data = _mutate_checkout_add_promo_code(api_client, variables)
@@ -1077,9 +1147,9 @@ def test_checkout_add_shipping_voucher_do_not_invalidate_shipping_method(
 
 
 MUTATION_CHECKOUT_REMOVE_PROMO_CODE = """
-    mutation($token: UUID, $promoCode: String, $promoCodeId: ID) {
+    mutation($id: ID, $promoCode: String, $promoCodeId: ID) {
         checkoutRemovePromoCode(
-            token: $token, promoCode: $promoCode, promoCodeId: $promoCodeId) {
+            id: $id, promoCode: $promoCode, promoCodeId: $promoCodeId) {
             errors {
                 field
                 code
@@ -1109,7 +1179,7 @@ def test_checkout_remove_voucher_code(api_client, checkout_with_voucher):
     previous_checkout_last_change = checkout_with_voucher.last_change
 
     variables = {
-        "token": checkout_with_voucher.token,
+        "id": to_global_id_or_none(checkout_with_voucher),
         "promoCode": checkout_with_voucher.voucher_code,
     }
 
@@ -1132,7 +1202,7 @@ def test_checkout_remove_voucher_code_with_inactive_channel(
     previous_checkout_last_change = checkout_with_voucher.last_change
 
     variables = {
-        "token": checkout_with_voucher.token,
+        "id": to_global_id_or_none(checkout_with_voucher),
         "promoCode": checkout_with_voucher.voucher_code,
     }
 
@@ -1150,7 +1220,7 @@ def test_checkout_remove_gift_card_code(api_client, checkout_with_gift_card):
     previous_checkout_last_change = checkout_with_gift_card.last_change
 
     variables = {
-        "token": checkout_with_gift_card.token,
+        "id": to_global_id_or_none(checkout_with_gift_card),
         "promoCode": checkout_with_gift_card.gift_cards.first().code,
     }
 
@@ -1173,7 +1243,7 @@ def test_checkout_remove_one_of_gift_cards(
     gift_card_last = checkout_with_gift_card.gift_cards.last()
 
     variables = {
-        "token": checkout_with_gift_card.token,
+        "id": to_global_id_or_none(checkout_with_gift_card),
         "promoCode": gift_card_first.code,
     }
 
@@ -1189,7 +1259,10 @@ def test_checkout_remove_one_of_gift_cards(
 
 def test_checkout_remove_promo_code_invalid_promo_code(api_client, checkout_with_item):
     previous_checkout_last_change = checkout_with_item.last_change
-    variables = {"token": checkout_with_item.token, "promoCode": "unexisting_code"}
+    variables = {
+        "id": to_global_id_or_none(checkout_with_item),
+        "promoCode": "unexisting_code",
+    }
 
     data = _mutate_checkout_remove_promo_code(api_client, variables)
 
@@ -1199,13 +1272,14 @@ def test_checkout_remove_promo_code_invalid_promo_code(api_client, checkout_with
     assert checkout_with_item.last_change == previous_checkout_last_change
 
 
-def test_checkout_remove_promo_code_invalid_checkout(api_client, voucher):
-    variables = {"token": uuid.uuid4(), "promoCode": voucher.code}
+def test_checkout_remove_promo_code_invalid_checkout(api_client, voucher, checkout):
+    variables = {"id": to_global_id_or_none(checkout), "promoCode": voucher.code}
+    checkout.delete()
 
     data = _mutate_checkout_remove_promo_code(api_client, variables)
 
     assert data["errors"]
-    assert data["errors"][0]["field"] == "token"
+    assert data["errors"][0]["field"] == "id"
 
 
 def test_checkout_remove_voucher_code_by_id(
@@ -1215,7 +1289,7 @@ def test_checkout_remove_voucher_code_by_id(
     checkout_with_voucher.gift_cards.add(gift_card)
 
     variables = {
-        "token": checkout_with_voucher.token,
+        "id": to_global_id_or_none(checkout_with_voucher),
         "promoCodeId": graphene.Node.to_global_id("Voucher", voucher.id),
     }
 
@@ -1236,7 +1310,7 @@ def test_checkout_remove_gift_card_by_id(
     checkout_with_voucher.gift_cards.add(gift_card, gift_card_expiry_date)
 
     variables = {
-        "token": checkout_with_voucher.token,
+        "id": to_global_id_or_none(checkout_with_voucher),
         "promoCodeId": graphene.Node.to_global_id("GiftCard", gift_card.id),
     }