+++ saleor/graphql/order/enums.py
@@ -3,6 +3,8 @@ import graphene
 from ...graphql.core.enums import to_enum
 from ...order import (
     FulfillmentStatus,
+    OrderAuthorizeStatus,
+    OrderChargeStatus,
     OrderEvents,
     OrderEventsEmails,
     OrderOrigin,
@@ -14,6 +16,12 @@ OrderEventsEnum = to_enum(OrderEvents)
 OrderEventsEmailsEnum = to_enum(OrderEventsEmails)
 OrderOriginEnum = to_enum(OrderOrigin)
 OrderStatusEnum = to_enum(OrderStatus, type_name="OrderStatus")
+OrderAuthorizeStatusEnum = to_enum(
+    OrderAuthorizeStatus, description=OrderAuthorizeStatus.__doc__
+)
+OrderChargeStatusEnum = to_enum(
+    OrderChargeStatus, description=OrderChargeStatus.__doc__
+)
 
 
 class OrderStatusFilter(graphene.Enum):